Edit Categories
Admin now can edit categories

+++ Assignment/src/java/controller/CategoryDelete.java
@@ -2,7 +2,6 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/Servlet.java to edit this template
  */
-
 package controller;
 
 import dal.CategoryDAO;
@@ -12,33 +11,48 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import java.io.File;
+import java.nio.file.Paths;
 
 /**
  *
  * @author VHC
  */
 public class CategoryDelete extends HttpServlet {
-   
-    /** 
-     * Processes requests for both HTTP <code>GET</code> and <code>POST</code> methods.
+
+    /**
+     * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
+     * methods.
+     *
      * @param request servlet request
      * @param response servlet response
      * @throws ServletException if a servlet-specific error occurs
      * @throws IOException if an I/O error occurs
      */
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
+            throws ServletException, IOException {
         CategoryDAO DAO = new CategoryDAO();
-        
+
         String CategoryID = request.getParameter("CategoryId");
-        
-        DAO.deleteById(CategoryID);
-        request.getRequestDispatcher("/View/Home.jsp?Content=CategoryList.jsp").forward(request, response);
-    } 
+        String path = Paths.get(getServletContext().getRealPath("")).getParent().getParent().toString()
+                + File.separator + "web"
+                + File.separator + DAO.getCategory(CategoryID).getCategory_Img();
+
+        if (DelFile(path)) {
+            DAO.deleteById(CategoryID);
+        }
+        response.sendRedirect("View/Home.jsp?Content=CategoryList.jsp");
+    }
+
+    private boolean DelFile(String FilePath) {
+        File IMG = new File(FilePath);
+        return IMG.delete();
+    }
 
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
-    /** 
+    /**
      * Handles the HTTP <code>GET</code> method.
+     *
      * @param request servlet request
      * @param response servlet response
      * @throws ServletException if a servlet-specific error occurs
@@ -46,12 +60,13 @@ public class CategoryDelete extends HttpServlet {
      */
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
+            throws ServletException, IOException {
         processRequest(request, response);
-    } 
+    }
 
-    /** 
+    /**
      * Handles the HTTP <code>POST</code> method.
+     *
      * @param request servlet request
      * @param response servlet response
      * @throws ServletException if a servlet-specific error occurs
@@ -59,12 +74,13 @@ public class CategoryDelete extends HttpServlet {
      */
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
+            throws ServletException, IOException {
         processRequest(request, response);
     }
 
-    /** 
+    /**
      * Returns a short description of the servlet.
+     *
      * @return a String containing servlet description
      */
     @Override
